Change distribution to operating system in places visible to users
Distribution is a Linuxism that makes no sense for users of other
operating systems, and even many new Linux users are unfamiliar with
the term: some users try to add the name of a distributor or version to the distribution field. Operating system is more accurate and less likely to confuse  users.

Signed-off-by: Rosanne DiMesio <dimesio@earthlink.net>
Signed-off-by: Jeremy Newman <jnewman@codeweavers.com>