Add hasProblems yes/no question to the test report form
Adds question to form and checks user responses against the
responses to the Installs?, Runs?, and usedWorkarounds questions
and the rating given. The user is shown an error message if the
response/rating combination does not conform to the rating
definitions, e.g., rating cannot be Platinum if the answer
to this question is yes.

Signed-off-by: Rosanne DiMesio <dimesio@earthlink.net>
Signed-off-by: Jeremy Newman <jnewman@codeweavers.com>