Disable staging checkbox for new test reports
Prevents users from submitting new test reports for nonexistant staging versions.
Current data is retained and the checkbox is still active for existing reports
in case they need to be edited by an admin or maintainer. This change can
be easily reverted if the branch is officially revived by WineHQ.

Signed-off-by: Rosanne DiMesio <dimesio@earthlink.net>
Signed-off-by: Jeremy Newman <jnewman@codeweavers.com>